refactor(desktop): isolate scroll diagnostic contracts
Problem
Repository lint rejected new complexity in the transcript arbiter, bridge contract, and diagnostic payload validators.

Root cause
Diagnostic event mapping and recovery types lived in the single scroll-writer module, while manifest, summary, event-index, and label validation were combined into two large Go functions.

Fix
Extract pure diagnostic mapping, recovery types, and the additive Wails binding contract into focused modules. Split the Go validation boundary by payload responsibility without changing accepted fields or runtime behavior.

Verification
- go run ./tools/repolint
- pnpm test:transcript
- pnpm test:transcript-browser
- env -u REASONIX_CHANNEL pnpm build
- REASONIX_CHANNEL=test pnpm build
- cd desktop && go test ./...
- cd desktop && go vet ./...
