+++ desktop/frontend/src/lib/scrollDiagnosticBridge.ts
@@ -1,0 +1,3 @@
+export interface ScrollDiagnosticBindings {
+  ExportScrollDiagnostics?(payload: string): Promise<string>;
+}
